requirements.txt
sendgrid
requests
python-dotenv